Conflicts:
	templates/view_event.php

+++ templates/view_event.php
@@ -5,6 +5,9 @@ require_once (INCLUDES_PATH . "/authentication.php");
 require_once (INCLUDES_PATH . "/events.php");
 require_once (DATABASE_PATH . "/events.php");
 require_once (INCLUDES_PATH . "/utils.php");
+require_once (DATABASE_PATH . "/comment.php");
+
+require (INCLUDES_PATH . "/write_comment_action.php");
 
 try {
 	if (! isset ( $_GET ["id"] )) {
@@ -19,53 +22,57 @@ try {
 	} else {
 		$idEvent = $_GET ["id"];
 		$event = getEvent ( $idEvent );
+		$canEdit = (isUserLoggedIn () && $event ["owner"] === getUserID ());
+		
 		echo '<div class="event" id="event' . $idEvent . '">';
-		echo '<a href="delete_event.php?id=' . $idEvent . '" class="delete" id="delete_event"><img src="" alt="Delete Event" /></a>';
-		$public = $event ["public"];
-		if( $public )
-			echo '<a class="change_privacy public" id="change_privacy"><img src="" alt="Change Event Privacy" /><p class="description">Make me private</p></a>';
-		else
-			echo '<a class="change_privacy private" id="change_privacy"><img src="" alt="Change Event Privacy" /><p class="description">Make me public</p></a>';
+		if ($canEdit) {
+			echo '<a href="delete_event.php?id=' . $idEvent . '" class="delete" id="delete_event"><img src="" alt="Delete Event" /></a>';
+			$public = $event ["public"];
+			if( $public )
+				echo '<a class="change_privacy public" id="change_privacy"><img src="" alt="Change Event Privacy" /><p class="description">Make me private</p></a>';
+			else
+				echo '<a class="change_privacy private" id="change_privacy"><img src="" alt="Change Event Privacy" /><p class="description">Make me public</p></a>';
+		}
 		echo '<h1 id="name">' . htmlspecialchars ( $event ["name"] ) . '</h1>';
-		echo '<a href="" class="edit" id="edit_name"><img src="images/edit_field.png" alt="Edit" /></a>';
 		$eventTypeName = getEventTypeName($event ["type"]);
 		echo '<h2 id="type">' . $eventTypeName . '</h2>';
+		if ($canEdit)
+			echo '<a href="" class="edit" id="edit_name"><img src="images/edit_field.png" alt="Edit" /></a>';
 		echo '<div class="container">';
 		echo '<img id="image" src="database/event_image.php?id=' . $idEvent . '" alt="' . htmlspecialchars ( $event ["name"] ) . '" width="256" height="256" />';
 		echo '<p id="description">' . preg_replace ( "/\n/", "<br />", htmlspecialchars ( $event ["description"] ) ) . '</p>';
-		echo '<a href="" class="edit" id="edit_description"><img src="images/edit_field.png" alt="Edit" /></a>';
+		if ($canEdit)
+			echo '<a href="" class="edit" id="edit_description"><img src="images/edit_field.png" alt="Edit" /></a>';
 		echo '</div>';
+		
 		echo '<datetime id="date">' . htmlspecialchars ( $event ["date"] ) . '</datetime>';
-		echo '<a href="" class="edit" id="edit_date"><img src="images/edit_field.png" alt="Edit" /></a>';
+		if ($canEdit)
+			echo '<a href="" class="edit" id="edit_date"><img src="images/edit_field.png" alt="Edit" /></a>';
 		?>
-<div class="fb-share-button"
-	data-href="<?php echo requestedURL(); ?>"
+<div class="fb-share-button" data-href="<?php echo requestedURL(); ?>"
 	data-layout="button_count"></div>
 <?php
 		
 		echo '<div class="comment_area">';
-		echo '<form class="write_comment_form" id="write_comment" action="login.php" method="post">';
+		echo '<form class="write_comment_form" id="write_comment" action="view_event.php?id=' . $idEvent . '" method="post">';
+		echo '<input type="hidden" name="idEvent" value="' . $idEvent . '" />';
 		echo '<textarea name="text" id="text" required placeholder="Comment..." maxlength="500"></textarea>';
-		echo '<button id="submit" type="submit">Add comment</button>';
+		echo '<button id="submit" type="submit" name="submit_comment">Add comment</button>';
 		echo '</form>';
 		
 		echo '<div class="comment_container">';
 		echo '<h2 id="title">Comments:</h2>';
-		echo '<div class="comment">';
-		echo '<h3 id="user">Cristiano Ronaldo</h3>';
-		echo '<p id="text">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec efficitur, libero ac faucibus tincidunt, urna urna tincidunt enim, sit amet congue nunc magna at est. Nam rhoncus dignissim orci eget fermentum. Proin ultrices dignissim vestibulum. Suspendisse porttitor pellentesque suscipit. Fusce sodales, nisl et pretium rutrum, urna nunc egestas nulla, vitae imperdiet erat diam sit amet metus. Suspendisse po</p>';
-		echo '<h4 id="time">2014/02/02 22:22</h4>';
-		echo '</div>';
-		echo '<div class="comment">';
-		echo '<h3 id="user">Cristiano Ronaldo</h3>';
-		echo '<p id="text">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec efficitur, libero ac faucibus tincidunt, urna urna tincidunt enim, sit amet congue nunc magna at est. Nam rhoncus dignissim orci eget fermentum. Proin ultrices dignissim vestibulum. Suspendisse porttitor pellentesque suscipit. Fusce sodales, nisl et pretium rutrum, urna nunc egestas nulla, vitae imperdiet erat diam sit amet metus. Suspendisse po</p>';
-		echo '<h4 id="time">2014/02/02 22:22</h4>';
-		echo '</div>';
-		echo '<div class="comment">';
-		echo '<h3 id="user">Cristiano Ronaldo</h3>';
-		echo '<p id="text">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec efficitur, libero ac faucibus tincidunt, urna urna tincidunt enim, sit amet congue nunc magna at est. Nam rhoncus dignissim orci eget fermentum. Proin ultrices dignissim vestibulum. Suspendisse porttitor pellentesque suscipit. Fusce sodales, nisl et pretium rutrum, urna nunc egestas nulla, vitae imperdiet erat diam sit amet metus. Suspendisse po</p>';
-		echo '<h4 id="time">2014/02/02 22:22</h4>';
-		echo '</div>';
+		
+		$comments = getComments ( $idEvent );
+		
+		foreach ( $comments as $comment ) {
+			echo '<div class="comment">';
+			echo '<h3 id="user">' . htmlspecialchars ( $comment ["name"] ) . '</h3>';
+			echo '<p id="text">' . nl2br ( htmlspecialchars ( $comment ["text"] ) ) . '</p>';
+			echo '<h4 id="time">' . $comment ["date"] . '</h4>';
+			echo '</div>';
+		}
+		
 		echo '</div>';
 		echo '</div>';
 		echo '</div>';
